moved Filter folder to adhere to new convention and added extra padding to make room for focus state in optionDisplay
