+++ src/components/Filter/Filter.module.css
@@ -1,4 +1,4 @@
-@import "@digdir/design-system-tokens/dist/digdir/tokens.css";
+@import "@digdir/design-system-tokens/brand/digdir/tokens.css";
 
 .popoverContent{
     box-sizing: border-box;
