+++ src/components/Filter/OptionDisplay.module.css
@@ -1,5 +1,3 @@
-@import "@digdir/design-system-tokens/dist/digdir/tokens.css";
-
 .optionDisplay {
     display: flex;
     justify-content: center;
@@ -19,7 +17,7 @@
 .option {
     border: none;
     border-bottom: 1px solid var(--fds-semantic-border-neutral-divider);
-    padding: 6px 10px 6px 10px;
+    padding: 8px 10px 8px 10px;
     background-color: var(--fds-semantic-background-default);
     cursor: pointer;
     font: inherit;
